+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - Added **Tweets Manage** endpoints. ([#7](https://github.com/twitter-dart/twitter-api-v2/issues/7))
   - DELETE /2/tweets/:id
   - POST /2/tweets
+- Added **Tweets Retweet** endpoints. ([#9](https://github.com/twitter-dart/twitter-api-v2/issues/9))
+  - DELETE /2/users/:id/retweets/:source_tweet_id
+  - GET /2/tweets/:id/retweeted_by
+  - POST /2/users/:id/retweets
 
 ## v0.0.1